refactor: replaced font with typography
DSY-1456
