+++ src/common/themeSelectors/index.ts
@@ -2,9 +2,9 @@ export * from './avatarSizes/avatarSizes';
 export * from './button/button';
 export * from './colors/colors';
 export * from './effects/effects';
-export * from './fonts/fonts';
 export * from './opacity/opacity';
 export * from './sizes/sizes';
 export * from './spacing/spacing';
 export * from './helpers/helpers';
 export * from './theme/theme';
+export * from './typography/typography';
